requirements.txt
airsim
numpy
scipy
matplotlib
pyyaml
opencv-python
